requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aioice==0.10.2
aiortc==1.14.0
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
altair==6.0.0
annotated-types==0.7.0
asn1crypto==1.5.1
attrs==25.4.0
av==16.0.1
blinker==1.9.0
boto3==1.42.14
botocore==1.42.14
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
Cython==3.2.3
dnspython==2.8.0
easydict==1.13
filelock==3.20.1
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
google-crc32c==1.8.0
humanfriendly==10.0
idna==3.11
ifaddr==0.2.0
ImageIO==2.37.2
insightface==0.7.3
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib==3.10.8
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.7.0
narwhals==2.14.0
networkx==3.6.1
numpy==2.2.6
onnx==1.20.0
onnxruntime==1.23.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.1
prettytable==3.17.0
propcache==0.4.1
protobuf==6.33.2
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyee==13.0.0
PyJWT==2.10.1
pylibsrtp==1.0.0
pyOpenSSL==25.3.0
pyparsing==3.2.5
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
s3transfer==0.16.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.16.3
simsimd==6.5.12
six==1.17.0
smmap==5.0.2
snowflake-connector-python==4.1.1
snowflake-snowpark-python==1.44.0
sortedcontainers==2.4.0
streamlit==1.52.2
streamlit-webrtc==0.64.5
stringzilla==4.5.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.12.20
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.4
tqdm==4.67.1
twilio==9.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
watchdog==6.0.0
wcwidth==0.2.14
yarl==1.22.0
